Remove double newline on debug
When debug mode is active and a file properly terminated
with a newline is included, an extra empty line appears before
the "End:" closing comment.

With this patch, no extra blank line will be inserted.

If the included file does not terminate with an empty line, the
debug "End:" comment will appear just after the included content,
which is both appropriate (to show that there isn't a newline) and 
harmless.